refactor file structure and move util functions
