+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2024-10-03
+
+### ADDED
+
+- Zip compression of output data folder to save storage space
+
+### CHANGED
+
+- Renamed naming of "Schema" to "Recipe" to reduce confusion, since recipe contains both schema and derivatives
+- Output of file from `/data` to `/output`
+- Global nullable percentage from `0.2` to `0.1`
+
 ## [1.0.0] - 2024-09-26
 
 ### ADDED
